requirements.txt
ipython==7.18.1
redis==3.5.3
TA-lib==0.4.18
pyfolio==0.9.2
#zipline==1.4.0
matplotlib==3.3.2
nose==1.3.7
numpy==1.19.2
pandas==1.1.2
scipy==1.5.2
requests==2.24.0
dateparser==0.7.6
environs==8.0.0
git+https://github.com/Telstra/MessagingAPI-SDK-python.git
#python-binance-chain==0.1.20
python-binance==0.7.5
python-telegram-bot==12.8